Update AssetLibPreview.cs
Added Asset ID to help assist in identifying correct Addon in Preview.
Added check to see if signal download_complete has been connected, and connect as a OneShot.
Added check to see if Meta has been set for iconPath, if it hasn't, set the Missing Icon, otherwise go through the process of trying to load the icon.