Add Import files
Per Issue godotengine/godot#72273 in comment from akien-mga, import files should be committed to Github repo, as it contains import settings for textures, and other files, as well as the associated UID for said files.  Another commit will come with the scenes being committed as well for the resaves to new UID's assigned to files.