Update ImportProject.cs
Updated to show warning when there's a Older version then the currently selected option.  Need to verify how to do Update from 3.x to 4.x, if it's opened with Godot 4, but warning for now to show that there is something to worry about when setting the Godot version to a higher level.