Major Refactor.
Removed Tuxfamily Download as option.  Godot developers are considering deprecation of the mirror in the future, and may not be a reliable source.  May re-implement it in the future, for now, it has been removed.

Upgraded engine/framework to 4.2.

Replaced all FileDialogs/BrowseFolderDialogs with Godot's new implementation of Native File Dialogs.

Removed C# Library for NFD.

Updated Mirrors to use Official Release, and Developer Builds.

Need to implement new Manager for Builds for Github manager.