Added Logic to hold Currently Selected Line Item and Icon Item, for later use with Delete icon in Actions bar.
Added handling of Creating and Removing Categories.

Added logic to properly refresh Category view when new categories are added.

Added logic to handle Selecting of Line Item, and Icon Item.

Added catch all, if a Category isn't null, but the Category doesn't exist, we set the category to be null, to prevent Null Exceptions.