Fixes #93
Stops attempting to launch the Editor when the Godot Version is null.
Fixed crash that when attempting to edit the project with a null version.
Added Check for null version for the engine when attempting to edit the project, which would crash before, will no longer crash, and allow the user to select the version of Godot to associate with the project.
If we have only 1 version of Godot Engine installed, and no associated engine with the project, ensure that the project is marked dirty, and enable the save button, allowing for association of a project with an Engine Version.