Fixes #98
Due to how the API is designed, two concurrent requests for Mirrors were being executed at the same time, causing one to not be completed, while the other one was, causing either the Setup Wizard, or the Engine panel not to properly populate, and ensure release type / standard/c# menus to be properly setup.  With this, we add a static bool toggle, for when we are requesting mirrors, and await IdleFrames till the request has been completed, before continuing, allowing for these simulanous calls to be made, without messing up any of the coding.