Updated NewsPanel
Created Callable to update Byline, and use CallDeferred() to ensure that Updates to the BusyDialog is run on the Main Thread.
Fixed Author Name, to ensure that any " or ' are removed from the Author Name, that can cause mis-matches.
Cleaned up Author Fetching, will attempt to fetch the Author, if not found, then attempt to fetch the default icon, and if for some reason, there is no default, then we just add the news item, and exit out.
Added Else statement to checking for pre-existing image downloader for the Avatar Image.  If the downloader is found, then we add a new Event hook, to update the author's image when the download is completed.  This will ensure, multiple news posts from the same author, are properly updated.
Removed code for UpdateQueue to debug code.  Now only prints errors when the Download failed, or Cancelled, and uses missing_icon.svg for the image.