Fixed Bug with Author Avatars
Fixed error, where newsItem.Avatar was being set in the wrong place, it should be set if the File does exist, not inside the queue check for multiple downloaders.
Fixed code for Multiple Downloaders of Avatar Images.  If a Downloader exists in the queue for the image in question, instead of creating a new one, we re-use the current one already created, and update the Meta data for the News Item to match what is set for the initial Image Downloader Creation.
Finally, Removed the break for when we are iterating through News Items, and we are matching AvatarDld, so that we can set multiple News Items, that are authored by the same Author.