Updated FileUtil
Commented out GetExtension(), since Godot has it's own GetExtension.
Added WildcardExists() WildcardFilename() for Asset Library Images, so that we can verify that the icons/previews are downloaded, without having to mess with images that do not have extensions, or crazy formatted filenames.
Added HasExtension() to see if an extension is provided or not, and use that instead.