Updated ImageDownloader
Updated to have OutputPath property.  When the downloader was created, it was being used for News, but now we are using it as well with Asset Library, so we ensure that we use it, if it was set when it was created.
Instead of just writing to the file, we test to see if we have an extension, if we don't, we attempt to detect the extension from the format of the image.