Updated SplashScreen
Updated SplashScreen to handle loading in a different method.  Before,
we would load the MainWindow scene by using the ChangeSceneToFile()
method, but this would cause delay as the Main Window was being
initialized (EG: Loading Author data, Image Data, fetching news, etc,
etc), so instead, we create the Main Window, and add it to the Root
hidden to start, and wait for the Ready event handler to be triggered,
before making the Main Window scene visible, and QueueFree'ing the
Splash Screen, making it more effective for loading process.